requirements.txt
# Black-Box Web Intelligence - Dependencies
# Core Orchestration
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-core>=0.3.0

# Browser Automation
playwright>=1.48.0
playwright-stealth>=1.0.6

# Schema & Validation
genson>=1.3.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Vector Store
chromadb>=0.5.0

# API Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
python-multipart>=0.0.12

# HTTP Client (for Verifier probes)
httpx>=0.27.0

# Database
aiosqlite>=0.20.0

# Utilities
simhash>=2.1.2
python-dotenv>=1.0.0
pyyaml>=6.0.2

# Async Support
asyncio>=3.4.3
aiofiles>=24.1.0
